refactor: move token amount parsing out of framework into token domain
