+++ src/features/transfer/utils/sponsoredSendExecution.ts
@@ -8,7 +8,7 @@ import { requireAddress } from '@/features/address/core/requireAddress';
 import { isPreparedCallsExecutionSponsored } from '@/features/delegation/utils/calls';
 import { type ChainId } from '@/features/network/types/backendNetworks';
 import { encodeErc20Transfer } from '@/features/token/core/services/erc20Calldata';
-import { parsePositiveRawAmount } from '@/framework/core/utils/units';
+import { parsePositiveRawAmount } from '@/features/token/core/services/tokenAmount';
 import { isNativeAsset } from '@/handlers/assets';
 import { resolveNameOrAddress } from '@/handlers/web3';
 import { type Call, type ExecuteCallsResult, type PreparedCallsExecution } from '@rainbow-me/sdk';
